[Feature]Refine module import. Fixes #69
