+++ examples/vizdoom/impala.py
@@ -12,12 +12,12 @@ from launchpad.nodes.python.local_multi_processing import PythonProcess
 
 from examples.vizdoom.network import network_maker
 from examples.vizdoom.utils import vizdoom_env_maker
-from moss.actor.vector import VectorActor
+from moss.actor import VectorActor
 from moss.agent.vizdoom import DoomAgent
-from moss.buffer.queue import QueueBuffer
+from moss.buffer import QueueBuffer
 from moss.env import EnvpoolVectorEnv, TimeStep
-from moss.learner.impala import ImpalaLearner
-from moss.predictor.base import BasePredictor
+from moss.learner import ImpalaLearner
+from moss.predictor import BasePredictor
 from moss.utils.loggers import Logger, experiment_logger_factory
 from moss.utils.paths import get_unique_id
