+++ moss/actor/__init__.py
@@ -1,1 +1,4 @@
 """Actor."""
+from moss.actor.vector import VectorActor
+
+__all__ = ["VectorActor"]
