+++ moss/agent/__init__.py
@@ -1,1 +1,4 @@
 """Agent."""
+from moss.agent.base import BaseAgent
+
+__all__ = ["BaseAgent"]
