+++ moss/buffer/__init__.py
@@ -1,1 +1,4 @@
 """Replay buffer."""
+from moss.buffer.queue import QueueBuffer
+
+__all__ = ["QueueBuffer"]
