+++ moss/learner/__init__.py
@@ -1,1 +1,6 @@
 """Learner."""
+from moss.learner.base import BaseLearner
+from moss.learner.impala import ImpalaLearner
+from moss.learner.ppo import PPOLearner
+
+__all__ = ["BaseLearner", "ImpalaLearner", "PPOLearner"]
