+++ moss/network/__init__.py
@@ -1,1 +1,6 @@
 """Network."""
+from moss.network.base import CommonNet, Network
+from moss.network.torso import DenseTorso
+from moss.network.value import DenseValue
+
+__all__ = ["CommonNet", "DenseTorso", "DenseValue", "Network"]
