+++ moss/predictor/__init__.py
@@ -1,1 +1,4 @@
 """Predicotr."""
+from moss.predictor.base import BasePredictor
+
+__all__ = ["BasePredictor"]
